Remove ethrex_l2 cli
