+++ cmd/ethrex/Cargo.toml
@@ -38,6 +38,7 @@ lazy_static.workspace = true
 secp256k1 = { workspace = true }
 keccak-hash.workspace = true
 reqwest.workspace = true
+itertools = "0.14.0"
 
 cfg-if = "1.0.0"
