+++ cmd/ethrex/l2/command.rs
@@ -9,17 +9,29 @@ use crate::{
     DEFAULT_L2_DATADIR,
 };
 use clap::Subcommand;
-use ethrex_common::{Address, U256};
-use ethrex_l2::SequencerConfig;
+use ethrex_common::{
+    types::{bytes_from_blob, BlockHeader, BYTES_PER_BLOB},
+    Address, U256,
+};
+use ethrex_l2::{sequencer::state_diff::StateDiff, SequencerConfig};
 use ethrex_p2p::network::peer_table;
 use ethrex_rpc::{
     clients::{beacon::BeaconClient, eth::BlockByNumber},
     EthClient,
 };
+use ethrex_storage::{EngineType, Store};
+use ethrex_storage_rollup::{EngineTypeRollup, StoreRollup};
 use eyre::OptionExt;
+use itertools::Itertools;
 use keccak_hash::keccak;
 use reqwest::Url;
-use std::{fs::create_dir_all, future::IntoFuture, path::PathBuf, sync::Arc, time::Duration};
+use std::{
+    fs::{create_dir_all, read_dir},
+    future::IntoFuture,
+    path::PathBuf,
+    sync::Arc,
+    time::Duration,
+};
 use tokio::sync::Mutex;
 use tokio_util::task::TaskTracker;
 use tracing::info;
@@ -55,6 +67,17 @@ pub enum Command {
         l1_eth_rpc: Url,
         #[arg(short = 'b', long)]
         l1_beacon_rpc: Url,
+    },
+    #[command(about = "Reconstructs the L2 state from L1 blobs.")]
+    Reconstruct {
+        #[arg(short = 'g', long, help = "The genesis file for the L2 network.")]
+        genesis: PathBuf,
+        #[arg(short = 'b', long, help = "The directory to read the blobs from.")]
+        blobs_dir: PathBuf,
+        #[arg(short = 's', long, help = "The path to the store.")]
+        store_path: PathBuf,
+        #[arg(short = 'c', long, help = "Address of the L2 proposer coinbase")]
+        coinbase: Address,
     },
 }
 
@@ -248,6 +271,124 @@ impl Command {
 
                     current_block += U256::one();
                 }
+            }
+            Command::Reconstruct {
+                genesis,
+                blobs_dir,
+                store_path,
+                coinbase,
+            } => {
+                // Init stores
+                let store = Store::new_from_genesis(
+                    store_path.to_str().expect("Invalid store path"),
+                    EngineType::Libmdbx,
+                    genesis.to_str().expect("Invalid genesis path"),
+                )
+                .await?;
+                let rollup_store = StoreRollup::new(
+                    store_path
+                        .join("./rollup_store")
+                        .to_str()
+                        .expect("Invalid store path"),
+                    EngineTypeRollup::Libmdbx,
+                )?;
+                rollup_store
+                    .init()
+                    .await
+                    .expect("Failed to init rollup store");
+
+                // Get genesis
+                let genesis_header = store.get_block_header(0)?.expect("Genesis block not found");
+                let genesis_block_hash = genesis_header.hash();
+
+                let mut new_trie = store
+                    .state_trie(genesis_block_hash)?
+                    .expect("Cannot open state trie");
+
+                let mut last_block_number = 0;
+
+                // Iterate over each blob
+                let files: Vec<std::fs::DirEntry> = read_dir(blobs_dir)?.try_collect()?;
+                for (file_number, file) in files
+                    .into_iter()
+                    .sorted_by_key(|f| f.file_name())
+                    .enumerate()
+                {
+                    let batch_number = file_number as u64 + 1;
+                    let blob = std::fs::read(file.path())?;
+
+                    if blob.len() != BYTES_PER_BLOB {
+                        panic!("Invalid blob size");
+                    }
+
+                    // Decode state diff from blob
+                    let blob = bytes_from_blob(blob.into());
+                    let state_diff = StateDiff::decode(&blob)?;
+
+                    // Apply all account updates to trie
+                    let account_updates = state_diff.to_account_updates(&new_trie)?;
+                    new_trie = store
+                        .apply_account_updates_from_trie(new_trie, &account_updates)
+                        .await
+                        .expect("Error applying account updates");
+
+                    // Get withdrawal hashes
+                    let withdrawal_hashes = state_diff
+                        .withdrawal_logs
+                        .iter()
+                        .map(|w| {
+                            keccak_hash::keccak(
+                                [
+                                    w.address.as_bytes(),
+                                    &w.amount.to_big_endian(),
+                                    w.tx_hash.as_bytes(),
+                                ]
+                                .concat(),
+                            )
+                        })
+                        .collect();
+
+                    // Get the first block of the batch
+                    let first_block_number = last_block_number + 1;
+
+                    // Build the header of the last block.
+                    // Note that its state_root is the root of new_trie.
+                    let new_block = BlockHeader {
+                        coinbase,
+                        state_root: new_trie.hash().expect("Error committing state"),
+                        ..state_diff.last_header
+                    };
+
+                    // Store last block.
+                    let new_block_hash = new_block.hash();
+                    store
+                        .add_block_header(new_block_hash, new_block.clone())
+                        .await?;
+                    store
+                        .add_block_number(new_block_hash, state_diff.last_header.number)
+                        .await?;
+                    store
+                        .set_canonical_block(state_diff.last_header.number, new_block_hash)
+                        .await?;
+                    println!(
+                        "Stored last block of blob. Block {}. State root {}",
+                        new_block.number, new_block.state_root
+                    );
+
+                    last_block_number = new_block.number;
+
+                    // Store batch info in L2 storage
+                    rollup_store
+                        .store_batch(
+                            batch_number,
+                            first_block_number,
+                            last_block_number,
+                            withdrawal_hashes,
+                        )
+                        .await
+                        .expect("Error storing batch");
+                }
+                store.update_latest_block_number(last_block_number).await?;
             }
         }
         Ok(())
