--- cmd/ethrex_l2/Cargo.toml
@@ -1,42 +1,0 @@
-[package]
-name = "ethrex_l2_cli"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-ethrex-l2.workspace = true
-ethrex-sdk.workspace = true
-ethrex-common.workspace = true
-ethrex-blockchain = { workspace = true, features = ["l2"] }
-ethrex-rlp.workspace = true
-ethrex-rpc.workspace = true
-ethrex-storage = { workspace = true, features = ["libmdbx"] }
-ethrex-storage-rollup = { workspace = true, features = ["libmdbx"] }
-
-serde_json.workspace = true
-serde.workspace = true
-bytes.workspace = true
-ethereum-types.workspace = true
-tokio.workspace = true
-tracing-subscriber = { workspace = true, features = ["env-filter"] }
-tracing.workspace = true
-hex.workspace = true
-log = "0.4"
-clap.workspace = true
-clap_complete.workspace = true
-eyre.workspace = true
-dirs = "5.0.1"
-toml = "0.8.19"
-dialoguer = "0.11"
-colored = "2.1.0"
-spinoff = "0.8.0"
-itertools = "0.13.0"
-strum = "0.26.3"
-secp256k1.workspace = true
-keccak-hash.workspace = true
-reqwest.workspace = true
-sha2.workspace = true
-
-[[bin]]
-name = "ethrex_l2"
-path = "./src/main.rs"
