--- cmd/ethrex_l2/rust-toolchain.toml
@@ -1,2 +1,0 @@
-[toolchain]
-channel = "1.83.0"
