--- cmd/ethrex_l2/src/commands/mod.rs
@@ -1,6 +1,0 @@
-pub(crate) mod autocomplete;
-pub(crate) mod config;
-pub(crate) mod info;
-pub(crate) mod stack;
-pub(crate) mod utils;
-pub(crate) mod wallet;
