--- cmd/ethrex_l2/src/utils/config/default_values.rs
@@ -1,29 +1,0 @@
-use ethereum_types::{H160, H256};
-
-pub const DEFAULT_CONFIG_NAME: &str = "local";
-pub const DEFAULT_L1_RPC_URL: &str = "http://localhost:8545";
-pub const DEFAULT_L1_CHAIN_ID: u64 = 3151908;
-pub const DEFAULT_L2_RPC_URL: &str = "http://localhost:1729";
-pub const DEFAULT_L2_CHAIN_ID: u64 = 1729;
-pub const DEFAULT_L2_EXPLORER_URL: &str = "";
-pub const DEFAULT_L1_EXPLORER_URL: &str = "";
-//0x385c546456b6a603a1cfcaa9ec9494ba4832da08dd6bcf4de9a71e4a01b74924
-pub const DEFAULT_PRIVATE_KEY: H256 = H256([
-    0x38, 0x5c, 0x54, 0x64, 0x56, 0xb6, 0xa6, 0x03, 0xa1, 0xcf, 0xca, 0xa9, 0xec, 0x94, 0x94, 0xba,
-    0x48, 0x32, 0xda, 0x08, 0xdd, 0x6b, 0xcf, 0x4d, 0xe9, 0xa7, 0x1e, 0x4a, 0x01, 0xb7, 0x49, 0x24,
-]);
-// 0x3d1e15a1a55578f7c920884a9943b3b35d0d885b
-pub const DEFAULT_ADDRESS: H160 = H160([
-    0x3d, 0x1e, 0x15, 0xa1, 0xa5, 0x55, 0x78, 0xf7, 0xc9, 0x20, 0x88, 0x4a, 0x99, 0x43, 0xb3, 0xb3,
-    0x5d, 0x0d, 0x88, 0x5b,
-]);
-// 0xaf6b492a2e60f9a366f3da02b78b6305b130edc7
-pub const DEFAULT_CONTRACTS_COMMON_BRIDGE_ADDRESS: H160 = H160([
-    0xaf, 0x6b, 0x49, 0x2a, 0x2e, 0x60, 0xf9, 0xa3, 0x66, 0xf3, 0xda, 0x02, 0xb7, 0x8b, 0x63, 0x05,
-    0xb1, 0x30, 0xed, 0xc7,
-]);
-// 0xea6d04861106c1fb69176d49eeb8de6dd14a9cfe
-pub const DEFAULT_CONTRACTS_ON_CHAIN_PROPOSER_ADDRESS: H160 = H160([
-    0xea, 0x6d, 0x04, 0x86, 0x11, 0x06, 0xc1, 0xfb, 0x69, 0x17, 0x6d, 0x49, 0xee, 0xb8, 0xde, 0x6d,
-    0xd1, 0x4a, 0x9c, 0xfe,
-]);
