--- cmd/ethrex_l2/src/utils/messages.rs
@@ -1,19 +1,0 @@
-// `ethrex_l2 config` messages
-pub const CONFIG_OVERRIDE_PROMPT_MSG: &str = "Config already exists. Do you want to overwrite it?";
-pub const CONFIG_CREATE_PROMPT_MSG: &str = "This config does not exist. Do you want to create it?";
-pub const CONFIG_CREATE_NAME_PROMPT_MSG: &str = "What do you want to name the new config?";
-pub const CONFIG_EDIT_PROMPT_MSG: &str = "What config do you want to edit?";
-pub const CONFIG_SET_PROMPT_MSG: &str = "What config do you want to set?";
-pub const CONFIG_DELETE_PROMPT_MSG: &str = "Are you sure you want to delete this config?";
-pub const CONFIG_SELECTION_TO_DELETE_PROMPT_MSG: &str = "What config do you want to delete?";
-pub const CONFIG_TO_DISPLAY_PROMPT_MSG: &str = "What config do you want to see?";
-pub const L1_RPC_URL_PROMPT_MSG: &str = "L1 RPC URL";
-pub const L1_CHAIN_ID_PROMPT_MSG: &str = "L1 CHAIN ID";
-pub const L2_RPC_URL_PROMPT_MSG: &str = "L2 RPC URL";
-pub const L2_CHAIN_ID_PROMPT_MSG: &str = "L2 CHAIN ID";
-pub const L2_EXPLORER_URL_PROMPT_MSG: &str = "L2 Explorer URL";
-pub const L1_EXPLORER_URL_PROMPT_MSG: &str = "L1 Explorer URL";
-pub const PRIVATE_KEY_PROMPT_MSG: &str = "Private key";
-pub const ADDRESS_PROMPT_MSG: &str = "Address";
-pub const CONTRACTS_COMMON_BRIDGE_PROMPT_MSG: &str = "Bridge contract address";
-pub const CONTRACTS_ON_CHAIN_PROPOSER_PROMPT_MSG: &str = "OnChainProposer contract address";
