--- cmd/ethrex_l2/src/utils/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod config;
-pub(crate) mod messages;
